refactor: Improved clarity of DispatchAreas in frontend& tests.
